+++ lib/firebase/auth/controller/firebase_auth_controller_interface.dart
@@ -1,0 +1,7 @@
+import 'package:jameswidget/firebase/firebase_response.dart';
+
+abstract class FirebaseAuthControllerInterface{
+  Future<FirebaseResponse> signIn() async => FirebaseResponse(error: FirebaseResponseErrorDetail(errorCheck: FirebaseResponseError.Error, errorData: ""), result: "");
+  Future<FirebaseResponse> signUp() async => FirebaseResponse(error: FirebaseResponseErrorDetail(errorCheck: FirebaseResponseError.Error, errorData: ""), result: "");
+  void dispose(){}
+}