+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.0.1
 
-* FirebaseEMailSigninController & FirebaseEMailSigninButton (2021.10.14)
+* (2021.10.14) FirebaseEMailSigninController & FirebaseEMailSigninButton
 
+## 0.0.2
 
+* (2021.10.15) Firebase EMail Sign In & Up + Example
+
+## 0.0.3
+* (2021.10.15) FirebaseReponse Model